Split ItemComponent into CollectableComponent as well
