+++ Engine/Systems/ItemCollectionSystem.cs
@@ -3,32 +3,35 @@ using System;
 
 namespace AdventureGame.Engine
 {
-    public class ItemSystem : System
+    public class ItemCollectionSystem : System
     {
-        public ItemSystem()
+        public ItemCollectionSystem()
         {
             RequiredComponent<ItemComponent>();
+            RequiredComponent<CollectableComponent>();
         }
 
         public override void UpdateEntity(GameTime gameTime, Scene scene, Entity entity)
         {
             ItemComponent itemComponent = entity.GetComponent<ItemComponent>();
+            CollectableComponent collectableComponent = entity.GetComponent<CollectableComponent>();
             ColliderComponent colliderComponent = entity.GetComponent<ColliderComponent>();
             //TransformComponent transformComponent = entity.GetComponent<TransformComponent>();
 
             // Check if the item is active and has not been collected
-            if (itemComponent.IsActive && !itemComponent.HasBeenCollected)
+            if (collectableComponent.IsActive && !collectableComponent.HasBeenCollected)
             {
                 // Respond to a collision between the item and another entity
                 foreach (Entity otherEntity in colliderComponent.collidedEntities)
                 {
                     // Check if the item can be collected by the other entity
-                    if (itemComponent.CanCollect(otherEntity.Tags))
+                    if (collectableComponent.CanCollect(otherEntity.Tags))
                     {
                         // Test - delete or move the item entity
                         // Should this be broadcast as a message instead to be picked up by InventorySystem?
                         // E.g. broadcast("itemPickup / itemCollected", entity, otherEntity) scene,time?
 
+                        // MOVE component declaration to the top?
                         // Check that the other entity has an inventory to add the item to
                         InventoryComponent inventoryComponent = otherEntity.GetComponent<InventoryComponent>();
                         if (inventoryComponent != null)
@@ -47,7 +50,7 @@ namespace AdventureGame.Engine
                                 entity.GetComponent<TransformComponent>().position.X += 50;
                                 itemComponent.Item.Quantity = origQuantity;
 
-                                if (itemComponent.Item.HasHealth())
+                                if (itemComponent.Item.HasItemHealth())
                                 {
                                     Random random = new Random();
                                     itemComponent.Item.ItemHealth = random.Next(0, itemComponent.Item.MaxHealth);
